Make a generic groupable tree item class
