+++ src/api.d.ts
@@ -41,6 +41,7 @@ export interface GroupNodeConfiguration {
     readonly keyLabel?: string;
     readonly description?: string;
     readonly icon?: vscode.ThemeIcon;
+    readonly iconPath?: string | vscode.Uri | { light: string | vscode.Uri; dark: string | vscode.Uri } | vscode.ThemeIcon;
     readonly contextValue?: string;
 }
