+++ src/tree/AppResourceTreeItem.ts
@@ -4,16 +4,12 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { AzExtParentTreeItem, IActionContext, nonNullProp, TreeItemIconPath } from "@microsoft/vscode-azext-utils";
-import * as path from 'path';
 import { FileChangeType } from "vscode";
 import { AppResource, GroupableResource, GroupingConfig, GroupNodeConfiguration, ResolvedAppResourceBase } from "../api";
 import { ext } from "../extensionVariables";
-import { createGroupConfigFromResource } from "../utils/azureUtils";
-import { treeUtils } from "../utils/treeUtils";
+import { createGroupConfigFromResource, getIconPath } from "../utils/azureUtils";
 import { GroupTreeItemBase } from "./GroupTreeItemBase";
-import { LocationGroupTreeItem } from "./LocationGroupTreeItem";
 import { ResolvableTreeItemBase } from "./ResolvableTreeItemBase";
-import { ResourceTypeGroupTreeItem } from "./ResourceTypeGroupTreeItem";
 import { SubscriptionTreeItem } from "./SubscriptionTreeItem";
 
 export class AppResourceTreeItem extends ResolvableTreeItemBase implements GroupableResource {
@@ -33,12 +29,9 @@ export class AppResourceTreeItem extends ResolvableTreeItemBase implements Group
         this.rootGroupConfig = <GroupNodeConfiguration><unknown>parent;
 
         this.data = resource;
-        this.groupConfig = createGroupConfigFromResource(resource);
+        this.groupConfig = createGroupConfigFromResource(resource, parent.id);
 
         this.contextValues.add(AppResourceTreeItem.contextValue);
-
-        // test for [label: string] keys
-        this.groupConfig["location"] = { keyLabel: 'Locations', label: this.data.location || 'unknown', id: `${this.parent?.id}/${this.data.location}` }
         ext.tagFS.fireSoon({ type: FileChangeType.Changed, item: this });
     }
 
@@ -83,24 +76,7 @@ export class AppResourceTreeItem extends ResolvableTreeItemBase implements Group
     }
 
     public get iconPath(): TreeItemIconPath {
-        let iconName: string;
-        const rType: string | undefined = this.data.type?.toLowerCase();
-        if (rType && supportedIconTypes.includes(rType)) {
-            iconName = rType;
-            switch (rType) {
-                case 'microsoft.web/sites':
-                    if (this.data.kind?.toLowerCase().includes('functionapp')) {
-                        iconName = iconName.replace('sites', 'functionapp');
-                    }
-                    break;
-                default:
-            }
-            iconName = path.join('providers', iconName);
-        } else {
-            iconName = 'resource';
-        }
-
-        return treeUtils.getIconPath(iconName);
+        return getIconPath(this.data.type);
     }
 
     public async refreshImpl(): Promise<void> {
@@ -125,70 +101,14 @@ export class AppResourceTreeItem extends ResolvableTreeItemBase implements Group
         // const client = await createResourceClient([context, this.rootGroupTreeItem.subscription]);
         switch (groupBySetting) {
             case 'resourceType':
-                return new ResourceTypeGroupTreeItem(this.rootGroupTreeItem, this.groupConfig.resourceType.label)
+                return new GroupTreeItemBase(this.rootGroupTreeItem, this.groupConfig.resourceType);
             // case 'resourceGroup':
             // TODO: Use ResovableTreeItem here
             // return new ResourceGroupTreeItem(this.rootGroupTreeItem, (await client.resourceGroups.get(this.groupConfig.resourceGroup.label)));
             default:
-                return new LocationGroupTreeItem(this.rootGroupTreeItem, this.data.location?.toLocaleLowerCase() ?? 'No location');
+                return new GroupTreeItemBase(this.rootGroupTreeItem, this.groupConfig['location']);
         }
     }
 }
 
-// Execute `npm run listIcons` from root of repo to re-generate this list after adding an icon
-export const supportedIconTypes: string[] = [
-    'microsoft.web/functionapp',
-    'microsoft.web/hostingenvironments',
-    'microsoft.web/kubeenvironments',
-    'microsoft.web/serverfarms',
-    'microsoft.web/sites',
-    'microsoft.web/staticsites',
-    'microsoft.storage/storageaccounts',
-    'microsoft.sql/servers',
-    'microsoft.sql/servers/databases',
-    'microsoft.signalrservice/signalr',
-    'microsoft.servicefabricmesh/applications',
-    'microsoft.servicefabric/clusters',
-    'microsoft.servicebus/namespaces',
-    'microsoft.operationsmanagement/solutions',
-    'microsoft.operationalinsights/workspaces',
-    'microsoft.notificationhubs/namespaces',
-    'microsoft.network/applicationgateways',
-    'microsoft.network/applicationsecuritygroups',
-    'microsoft.network/loadbalancers',
-    'microsoft.network/localnetworkgateways',
-    'microsoft.network/networkinterfaces',
-    'microsoft.network/networksecuritygroups',
-    'microsoft.network/networkwatchers',
-    'microsoft.network/publicipaddresses',
-    'microsoft.network/publicipprefixes',
-    'microsoft.network/routetables',
-    'microsoft.network/virtualnetworkgateways',
-    'microsoft.network/virtualnetworks',
-    'microsoft.managedidentity/userassignedidentities',
-    'microsoft.logic/workflows',
-    'microsoft.kubernetes/connectedclusters',
-    'microsoft.keyvault/vaults',
-    'microsoft.insights/components',
-    'microsoft.extendedlocation/customlocations',
-    'microsoft.eventhub/namespaces',
-    'microsoft.eventgrid/domains',
-    'microsoft.eventgrid/eventsubscriptions',
-    'microsoft.eventgrid/topics',
-    'microsoft.documentdb/databaseaccounts',
-    'microsoft.devtestlab/labs',
-    'microsoft.devices/iothubs',
-    'microsoft.dbforpostgresql/servers',
-    'microsoft.dbformysql/servers',
-    'microsoft.containerservice/managedclusters',
-    'microsoft.containerregistry/registries',
-    'microsoft.compute/availabilitysets',
-    'microsoft.compute/disks',
-    'microsoft.compute/images',
-    'microsoft.compute/virtualmachines',
-    'microsoft.compute/virtualmachinescalesets',
-    'microsoft.cdn/profiles',
-    'microsoft.cache/redis',
-    'microsoft.batch/batchaccounts',
-    'microsoft.apimanagement/service',
-];
+
